Move conserver logic into Console
So that we can use it for vps nodes
